// src/generated/whatsapp_poll_types.ts
export interface PollChoice {
  index: number;
  text: string;
  voteCount: number;
}

export interface Poll {
  pollId: string;
  question: string;
  choices: PollChoice[];
  allowMultipleChoices: boolean;
  hideVoterNames: boolean;
}

export interface PollChangeEvent {
  recipient: string;
  pollId: string;
  occurredAt: Date | undefined;
  isFromMe: boolean;
  created?: Poll | undefined;
  updated?: Poll | undefined;
  voteChanged?: Poll | undefined;
  choicesChanged?: Poll | undefined;
}

function createBasePollChoice(): PollChoice {
  return { index: 0, text: "", voteCount: 0 };
}

export const PollChoice: MessageFns<PollChoice> = {
  encode(message: PollChoice, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.index !== 0) {
      writer.uint32(8).int32(message.index);
    }
    if (message.text !== "") {
      writer.uint32(18).string(message.text);
    }
    if (message.voteCount !== 0) {
      writer.uint32(24).int32(message.voteCount);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PollChoice {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePollChoice();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.index = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.text = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.voteCount = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PollChoice {
    return {
      index: isSet(object.index) ? globalThis.Number(object.index) : 0,
      text: isSet(object.text) ? globalThis.String(object.text) : "",
      voteCount: isSet(object.voteCount)
        ? globalThis.Number(object.voteCount)
        : isSet(object.vote_count)
        ? globalThis.Number(object.vote_count)
        : 0,
    };
  },

  toJSON(message: PollChoice): unknown {
    const obj: any = {};
    if (message.index !== 0) {
      obj.index = Math.round(message.index);
    }
    if (message.text !== "") {
      obj.text = message.text;
    }
    if (message.voteCount !== 0) {
      obj.voteCount = Math.round(message.voteCount);
    }
    return obj;
  },

  create(base?: DeepPartial<PollChoice>): PollChoice {
    return PollChoice.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<PollChoice>): PollChoice {
    const message = createBasePollChoice();
    message.index = object.index ?? 0;
    message.text = object.text ?? "";
    message.voteCount = object.voteCount ?? 0;
    return message;
  },
};

function createBasePoll(): Poll {
  return { pollId: "", question: "", choices: [], allowMultipleChoices: false, hideVoterNames: false };
}

export const Poll: MessageFns<Poll> = {
  encode(message: Poll, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.pollId !== "") {
      writer.uint32(10).string(message.pollId);
    }
    if (message.question !== "") {
      writer.uint32(18).string(message.question);
    }
    for (const v of message.choices) {
      PollChoice.encode(v!, writer.uint32(26).fork()).join();
    }
    if (message.allowMultipleChoices !== false) {
      writer.uint32(32).bool(message.allowMultipleChoices);
    }
    if (message.hideVoterNames !== false) {
      writer.uint32(40).bool(message.hideVoterNames);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Poll {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePoll();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.pollId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.question = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.choices.push(PollChoice.decode(reader, reader.uint32()));
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.allowMultipleChoices = reader.bool();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.hideVoterNames = reader.bool();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Poll {
    return {
      pollId: isSet(object.pollId)
        ? globalThis.String(object.pollId)
        : isSet(object.poll_id)
        ? globalThis.String(object.poll_id)
        : "",
      question: isSet(object.question) ? globalThis.String(object.question) : "",
      choices: globalThis.Array.isArray(object?.choices) ? object.choices.map((e: any) => PollChoice.fromJSON(e)) : [],
      allowMultipleChoices: isSet(object.allowMultipleChoices)
        ? globalThis.Boolean(object.allowMultipleChoices)
        : isSet(object.allow_multiple_choices)
        ? globalThis.Boolean(object.allow_multiple_choices)
        : false,
      hideVoterNames: isSet(object.hideVoterNames)
        ? globalThis.Boolean(object.hideVoterNames)
        : isSet(object.hide_voter_names)
        ? globalThis.Boolean(object.hide_voter_names)
        : false,
    };
  },

  toJSON(message: Poll): unknown {
    const obj: any = {};
    if (message.pollId !== "") {
      obj.pollId = message.pollId;
    }
    if (message.question !== "") {
      obj.question = message.question;
    }
    if (message.choices?.length) {
      obj.choices = message.choices.map((e) => PollChoice.toJSON(e));
    }
    if (message.allowMultipleChoices !== false) {
      obj.allowMultipleChoices = message.allowMultipleChoices;
    }
    if (message.hideVoterNames !== false) {
      obj.hideVoterNames = message.hideVoterNames;
    }
    return obj;
  },

  create(base?: DeepPartial<Poll>): Poll {
    return Poll.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<Poll>): Poll {
    const message = createBasePoll();
    message.pollId = object.pollId ?? "";
    message.question = object.question ?? "";
    message.choices = object.choices?.map((e) => PollChoice.fromPartial(e)) || [];
    message.allowMultipleChoices = object.allowMultipleChoices ?? false;
    message.hideVoterNames = object.hideVoterNames ?? false;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}